+++ packages/node/src/login/oidc/refresh/TokenRefresher.spec.ts
@@ -43,8 +43,10 @@ import {
   mockDefaultIssuerConfig,
   mockIssuerConfigFetcher,
 } from "../__mocks__/IssuerConfigFetcher";
+import { negotiateClientSigningAlg } from "../ClientRegistrar";
 
 jest.mock("openid-client");
+jest.mock("../ClientRegistrar");
 
 const mockJwk = (): JWK => {
   return {
@@ -119,22 +121,17 @@ const mockDpopTokens = (): TokenSet => {
 const setupOidcClientMock = (tokenSet: TokenSet) => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const { Issuer } = jest.requireMock("openid-client") as any;
-  function clientConstructor() {
-    // this is untyped, which makes TS complain
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.refresh = jest.fn().mockResolvedValueOnce(tokenSet);
-  }
   const mockedIssuer = {
     metadata: mockDefaultIssuerConfig(),
-    Client: clientConstructor,
+    Client: jest.fn().mockReturnValue({
+      refresh: jest.fn().mockResolvedValueOnce(tokenSet as never),
+    }),
   };
   Issuer.mockReturnValueOnce(mockedIssuer);
+  return mockedIssuer;
 };
 
-const setupDefaultOidcClientMock = () => {
-  setupOidcClientMock(mockDpopTokens());
-};
+const setupDefaultOidcClientMock = () => setupOidcClientMock(mockDpopTokens());
 
 const mockDefaultStorageContent = {
   "solidClientAuthenticationUser:mySession": {
@@ -206,6 +203,13 @@ describe("TokenRefresher", () => {
 
   it("throws if a refresh token isn't provided", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest
+      .fn(negotiateClientSigningAlg)
+      .mockReturnValue("ES256");
+
     const mockedStorage = mockRefresherDefaultStorageUtility();
 
     const refresher = getTokenRefresher({
@@ -219,6 +223,12 @@ describe("TokenRefresher", () => {
 
   it("throws if a DPoP token is expected, but no DPoP key is provided", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest
+      .fn(negotiateClientSigningAlg)
+      .mockReturnValue("ES256");
     const mockedStorage = mockRefresherDefaultStorageUtility();
 
     const refresher = getTokenRefresher({
@@ -232,10 +242,12 @@ describe("TokenRefresher", () => {
     );
   });
 
-  // FIXME: this test brings coverage to 100%, but has no meaningful expect.
-  // For the time being, it is sufficient, but to be fixed soon.
   it("does not negotiate the signing algorithm if it is already set for the client", async () => {
     setupDefaultOidcClientMock();
+    const mockedModule = jest.requireMock("../ClientRegistrar") as {
+      negotiateClientSigningAlg: typeof negotiateClientSigningAlg;
+    };
+    mockedModule.negotiateClientSigningAlg = jest.fn();
     const refresher = getTokenRefresher({
       storageUtility: mockRefresherDefaultStorageUtility(),
       clientRegistrar: mockClientRegistrar({
@@ -246,13 +258,36 @@ describe("TokenRefresher", () => {
       }),
     });
 
-    const refreshedTokens = await refresher.refresh(
+    await refresher.refresh(
       "mySession",
       "some refresh token",
       await mockKeyPair()
     );
 
-    expect(refreshedTokens.accessToken).toBe(mockDpopTokens().access_token);
+    expect(mockedModule.negotiateClientSigningAlg).not.toHaveBeenCalled();
+  });
+
+  it("uses 'none' authentication if using Solid-OIDC client identifiers", async () => {
+    const mockedIssuer = setupDefaultOidcClientMock();
+    const refresher = getTokenRefresher({
+      clientRegistrar: mockClientRegistrar({
+        clientId: "https://some.client.identifier",
+        clientType: "solid-oidc",
+        idTokenSignedResponseAlg: "ES256",
+      }),
+    });
+
+    await refresher.refresh(
+      "mySession",
+      "some refresh token",
+      await mockKeyPair()
+    );
+
+    expect(mockedIssuer.Client).toHaveBeenCalledWith(
+      expect.objectContaining({
+        token_endpoint_auth_method: "none",
+      })
+    );
   });
 
   it("refreshes a DPoP token properly", async () => {